requirements.txt
openpyxl
PyQt6pip